requirements.txt
alembic==1.13.2
bcrypt==4.2.0
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
PyJWT==2.9.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
Werkzeug==3.0.4
